% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_to_ppm.R
\name{abs_to_ppm}
\alias{abs_to_ppm}
\title{Calculate analyte concentration (ppm) from plate reader absorbance values}
\usage{
abs_to_ppm(assay.data.raw, sample.IDs, standard.IDs)
}
\arguments{
\item{assay.data.raw}{An excel table exported from the Gen5 software.}

\item{sample.IDs}{A 2-column dataframe consisting of 'well_ID' (e.g., A12) and 'sample' (a sample ID mapped to each well_ID)}

\item{standard.IDs}{A 2-column dataframe consisting of 'well_ID' (e.g., A12) and 'standard' (a standard concentration, generally in parts per million, mapped to each well_ID)}
}
\description{
This function calculates mean OTU or ASV count data based on a defined grouping_variable.
}
\examples{
abs_to_ppm
}
\keyword{calculate}
\keyword{concentration}
\keyword{excel}
