% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_baranyi_plot.R
\name{add_baranyi_plot}
\alias{add_baranyi_plot}
\title{Add Baranyi curves to a plot object}
\usage{
add_baranyi_plot(p, plate, baranyi)
}
\arguments{
\item{p}{A ggplot base object}

\item{plate}{A plate object/tibble}

\item{gompertz}{A tibble/df output from fit_baranyi_plate}
}
\description{
Add Baranyi curves to a plot object
}
