% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_biolog_metadata.R
\name{add_biolog_metadata}
\alias{add_biolog_metadata}
\title{Add biolog metadata}
\usage{
add_biolog_metadata(plate, type = "PM1")
}
\arguments{
\item{plate}{A plate object/tibble}

\item{type}{The biolog plate type (PM1, PM2A, etc)}
}
\description{
Add biolog metadata
}
