% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gompertz_plot.R
\name{add_gompertz_plot}
\alias{add_gompertz_plot}
\title{Add Gompertz curves to a plot object}
\usage{
add_gompertz_plot(p, plate, gompertz)
}
\arguments{
\item{p}{A ggplot base object}

\item{plate}{A plate object/tibble}

\item{gompertz}{A tibble/df output from fit_gompertz_plate}
}
\description{
Add Gompertz curves to a plot object
}
