% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logistic_plot.R
\name{add_logistic_plot}
\alias{add_logistic_plot}
\title{Add Logistic curves to a plot object}
\usage{
add_logistic_plot(p, plate, logistic)
}
\arguments{
\item{p}{A ggplot base object}

\item{plate}{A plate object/tibble}

\item{gompertz}{A tibble/df output from fit_logistic_plate}
}
\description{
Add Logistic curves to a plot object
}
