% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_substrate_to_plot.R
\name{add_substrate_to_plot}
\alias{add_substrate_to_plot}
\title{Add biolog substrates to a plot object}
\usage{
add_substrate_to_plot(p, plate)
}
\arguments{
\item{p}{A ggplot base object}

\item{plate}{A plate object/tibble that has already been run through add_biolog_metadata
#'}
}
\description{
Add biolog substrates to a plot object
}
