% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_well.R
\name{analyze_well}
\alias{analyze_well}
\title{Analyze Well}
\usage{
analyze_well(plate, well)
}
\arguments{
\item{plate}{A plate object/tibble}

\item{well}{A single well to inspect}
}
\description{
Analyze Well
}
