% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_subtract.R
\name{background_subtract}
\alias{background_subtract}
\title{Background subtract plate data}
\usage{
background_subtract(plate)
}
\arguments{
\item{plate}{A plate object/tibble}
}
\description{
Background subtract plate data
}
