% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biolog}
\alias{biolog}
\title{A dataset with biolog metadata}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 288 rows and 7 columns.}
\usage{
biolog
}
\description{
A dataset with biolog metadata
}
\keyword{datasets}
