% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biolog_plot_and_table.R
\name{biolog_plot_and_table}
\alias{biolog_plot_and_table}
\title{Convert plate data to plot and tibble}
\usage{
biolog_plot_and_table(x, w = "600")
}
\arguments{
\item{x}{A df output from read_multiplate_biolog_experiment()}

\item{w}{A wavelength (or vector of wavelengths)}
}
\description{
Convert plate data to plot and tibble
}
