% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_baranyi.R
\name{fit_baranyi}
\alias{fit_baranyi}
\title{Fit a single sample with Baranyi}
\usage{
fit_baranyi(x, y)
}
\arguments{
\item{x}{Time vector}

\item{y}{Growth vector}
}
\description{
Fit a single sample with Baranyi
}
