% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_baranyi_plate.R
\name{fit_baranyi_plate}
\alias{fit_baranyi_plate}
\title{Fit a plate with Baranyi}
\usage{
fit_baranyi_plate(plate)
}
\arguments{
\item{plate}{A plate object/tibble}
}
\description{
Fit a plate with Baranyi
}
