% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_biolog_substrate.R
\name{fit_biolog_substrate}
\alias{fit_biolog_substrate}
\title{Return fit of a substrate}
\usage{
fit_biolog_substrate(x, s = "Glucose", w = "600")
}
\arguments{
\item{x}{A raw biolog dataframe from read_multiplate_biolog_experiment()}

\item{s}{Substrate}

\item{w}{Measured wavelength}
}
\description{
Return fit of a substrate
}
