% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gompertz.R
\name{fit_gompertz}
\alias{fit_gompertz}
\title{Fit a single sample with gompertz}
\usage{
fit_gompertz(x, y)
}
\arguments{
\item{x}{Time vector}

\item{y}{Growth vector}
}
\description{
Fit a single sample with gompertz
}
