% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gompertz_plate.R
\name{fit_gompertz_plate}
\alias{fit_gompertz_plate}
\title{Fit a plate with Gompertz}
\usage{
fit_gompertz_plate(plate)
}
\arguments{
\item{plate}{A plate object/tibble}
}
\description{
Fit a plate with Gompertz
}
