% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_logistic.R
\name{fit_logistic}
\alias{fit_logistic}
\title{Fit a single sample with logistic equation}
\usage{
fit_logistic(x, y)
}
\arguments{
\item{x}{Time vector}

\item{y}{Growth vector}
}
\description{
Fit a single sample with logistic equation
}
