% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_logistic_plate.R
\name{fit_logistic_plate}
\alias{fit_logistic_plate}
\title{Fit a plate with Logistic}
\usage{
fit_logistic_plate(plate)
}
\arguments{
\item{plate}{A plate object/tibble}
}
\description{
Fit a plate with Logistic
}
