% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nonparametric.R
\name{fit_nonparametric}
\alias{fit_nonparametric}
\title{Fit a single sample with local non-parametric}
\usage{
fit_nonparametric(x, y, h = 4)
}
\arguments{
\item{x}{Time vector}

\item{y}{Growth vector}
}
\description{
Fit a single sample with local non-parametric
}
