% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthMetrics.R
\name{growthMetrics}
\alias{growthMetrics}
\title{Calculate growth metrics from a df}
\usage{
growthMetrics(x, wellRow = "A", t_min = 5, t_max = FALSE)
}
\arguments{
\item{x}{A df from functions such as read_time_series_columns.R}

\item{wellRow}{Row to calculate growth metrics for}

\item{t_min}{An beginning timepoint in case the lag phase is weird}

\item{t_max}{An endpoint time in case the fit isn't optimal due to wonky late stationary phase death}
}
\description{
Calculate growth metrics from a df
}
