% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_base_data.R
\name{plot_base_data}
\alias{plot_base_data}
\title{Make a ggplot object which can act as a base for further mappings}
\usage{
plot_base_data(plate, by_substrate = F)
}
\arguments{
\item{plate}{A plate object/tibble}

\item{by_substrate}{Organize the plate by substrate (TRUE) or well (FALSE, default)
#'}
}
\description{
Make a ggplot object which can act as a base for further mappings
}
