% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_grc_models.R
\name{plot_growth}
\alias{plot_growth}
\title{Run growthcurver models}
\usage{
plot_growth(x, line_col = "black", pt_col, plot_curves = TRUE, facets = TRUE)
}
\arguments{
\item{x}{A data.frame containing predictions from \code{run_grc_models}}

\item{line_col}{color of fitted lines}

\item{pt_col}{A factor in predictions data.frame to be used to determine colors
of data points.}

\item{plot_curves}{Logical. Whether to plot growth curves or not.}

\item{facets}{Logical. Whether to plot facets or not.}
}
\description{
Run growthcurver models
}
\author{
Ethan Bass
}
