% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_multiplate_biolog_experiment.R
\name{read_multiplate_biolog_experiment}
\alias{read_multiplate_biolog_experiment}
\title{Read Series of Excel Files in a Directory}
\usage{
read_multiplate_biolog_experiment(path = "", plate = "PM1")
}
\arguments{
\item{path}{A folder path containing files}
}
\description{
Read Series of Excel Files in a Directory
}
