% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_multiplate_experiment.R
\name{read_multiplate_experiment}
\alias{read_multiplate_experiment}
\title{Read Series of Excel Files in a Directory}
\usage{
read_multiplate_experiment(path = "")
}
\description{
Read Series of Excel Files in a Directory
}
