% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plate.R
\name{read_plate}
\alias{read_plate}
\title{Read plate in txt format generated by Biotek Kc4 software.}
\usage{
read_plate(files)
}
\arguments{
\item{files}{List of file paths}
}
\description{
Read plate in txt format generated by Biotek Kc4 software.
}
\author{
Ethan Bass
}
