% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plate_text.R
\name{read_plate_text}
\alias{read_plate_text}
\title{Read time series in a text file}
\usage{
read_plate_text(file = "")
}
\arguments{
\item{file}{An text file with wells as columns, and time points as rows}
}
\description{
Read time series in a text file
}
