% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_time_series_columns.R
\name{read_time_series_columns}
\alias{read_time_series_columns}
\title{Read time series in a single excel file}
\usage{
read_time_series_columns(file = "", repColumns = c(1, 2, 3),
  rowReplace = FALSE, wellRemove = "")
}
\arguments{
\item{file}{An excel file with wells as columns, and time points as rows}

\item{repColumns}{Columns for a replicate.}

\item{rowReplace}{Takes a mapping df with rows A-H, and assigns names to a new column}
}
\description{
Read time series in a single excel file
}
