% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_trim.R
\name{right_trim}
\alias{right_trim}
\title{Right Trim}
\usage{
right_trim(plate, timepoint)
}
\arguments{
\item{plate}{A plate object/tibble}

\item{timepoint}{Trim time points after this}
}
\description{
Right Trim
}
