% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_grc_models.R
\name{run_grc_model}
\alias{run_grc_model}
\title{Run growthcurver models}
\usage{
run_grc_model(x, lambda, what = c("models", "predictions"), plot_it = TRUE)
}
\arguments{
\item{x}{tidy growth curve data}

\item{lambda}{Character vector specifying wavelength (lambda) to use.}

\item{what}{what to return. Either \code{models}, \code{predictions}, or
\code{both}.}

\item{plot_it}{Whether to plot the growth curves}
}
\description{
Run growthcurver models
}
\note{
: This function is based on code by Daniel Padfield (https://github.com/sprouffske/growthcurver/issues/12)
}
\author{
Ethan Bass
}
