% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_blank.R
\name{subtract_blank}
\alias{subtract_blank}
\title{Subtract Blank}
\usage{
subtract_blank(plate, well)
}
\arguments{
\item{plate}{A plate object/tibble}

\item{well}{A single well to inspect}
}
\description{
Subtract Blank
}
